+++ kwcoco/util/__init__.py
@@ -1,0 +1,5 @@
+from kwcoco.util import util_futures
+
+from kwcoco.util.util_futures import (Executor, SerialExecutor, as_completed,)
+
+__all__ = ['Executor', 'SerialExecutor', 'as_completed', 'util_futures']