+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * Example usage section in CLI help
+* Moved toydata from ndsampler to kwcoco
 
 ### Fixed
 * Multipolygon segmentations are now displayed correctly.